db: No more pgx
Move to jmoiron/sqlx to interface with postgres. pgx was being really
annoying with scanning enums despite it just being text... and then had
issues scanning text into strings.

Current issues are that it's not liking the null value for
ephemeral_statuses by default, so let's see if we can fix that.